refactor: RootPageのファイル配置を他ページと統一
RootPage.tsx を RootPage/RootPage.tsx に移動し、
DetailPage, SettingsPage と同じディレクトリ構造に統一。
